Reimplemented zero/reset, changed getServoSpeed() to double

+++ Arduino/rover_arm_speed_test/joint.h
@@ -26,7 +26,7 @@ class joint
 
     // getters for debug
     double getSpeed();
-    int getServoSpeed();
+    double getServoSpeed();
 };
 
 #endif
